ISPN-561 Renamed jopr-plugin module to rhq-plugin
