+++ pom.xml
@@ -38,7 +38,7 @@
       <module>server/websocket</module>
       <module>server/rest</module>
       <module>client/hotrod-client</module>
-      <module>jopr-plugin</module>
+      <module>rhq-plugin</module>
       <module>demos/gui</module>
       <module>demos/ec2</module>
       <module>demos/ec2-ui</module>
