+++ rhq-plugin/pom.xml
@@ -13,11 +13,11 @@
       <relativePath>../parent/pom.xml</relativePath>
    </parent>
 
-   <artifactId>infinispan-jopr-plugin</artifactId>
+   <artifactId>infinispan-rhq-plugin</artifactId>
    <packaging>jar</packaging>
 
-   <name>Infinispan JOPR Plugin</name>
-   <description>The Infinispan JOPR Plugin</description>
+   <name>Infinispan RHQ Plugin</name>
+   <description>The Infinispan RHQ Plugin</description>
 
    <properties>
       <scm.module.path>plugin/infinispan</scm.module.path>
@@ -152,7 +152,7 @@
              <version>2.0</version>
              <executions>
                  <execution>
-                     <id>copy-infinispan-jopr-plugin-jar</id>
+                     <id>copy-infinispan-rhq-plugin-jar</id>
                      <phase>process-resources</phase>
                      <goals>
                          <goal>copy</goal>
@@ -176,7 +176,7 @@
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-javadoc-plugin</artifactId>
             <configuration>
-               <doclet>org.infinispan.tools.jopr.RhqPluginXmlGenerator</doclet>
+               <doclet>org.infinispan.tools.rhq.RhqPluginXmlGenerator</doclet>
                <docletArtifact>
                   <groupId>org.infinispan</groupId>
                   <artifactId>infinispan-tools</artifactId>
@@ -208,7 +208,7 @@
                </execution>
             </executions>
             <configuration>
-               <mainClass>org.infinispan.jopr.util.RhqPluginDescriptorGenerator</mainClass>
+               <mainClass>org.infinispan.rhq.util.RhqPluginDescriptorGenerator</mainClass>
             </configuration>
          </plugin>
          -->
