+++ rhq-plugin/src/main/java/org/infinispan/rhq/CacheManagerComponent.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.infinispan.jopr;
+package org.infinispan.rhq;
 
 import org.infinispan.util.logging.Log;
 import org.infinispan.util.logging.LogFactory;
