+++ src/main/resources/assemblies/bin.xml
@@ -133,7 +133,7 @@
             <include>org.infinispan:infinispan-cachestore-jdbm</include>
             <include>org.infinispan:infinispan-cachestore-s3</include>
             <include>org.infinispan:infinispan-gui-demo</include>
-            <include>org.infinispan:infinispan-jopr-plugin</include>
+            <include>org.infinispan:infinispan-rhq-plugin</include>
             <include>org.infinispan:infinispan-tree</include>
             <include>org.infinispan:infinispan-query</include>
             <include>org.infinispan:infinispan-lucene-directory</include>
